+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 **Table of Contents**  *generated with [DocToc](https://github.com/thlorenz/doctoc)*
 
+- [1.5.4](#154)
 - [1.5.3](#153)
 - [1.5.2](#152)
 - [1.5.1](#151)
@@ -36,6 +37,15 @@
 - [0.0.4](#004)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+
+## 1.5.4
+
+- Add `$setOnInsert` to modifiers for defaultValues only when `isUpsert` is set to `true` in clean options or in extended autoValue context. It used to be ignored but newer MongoDB versions throw an error. Might fix #304
+- Fix #307 - Test for empty object when creating schema (thanks @coagmano)
+- autoValue functions sort preserves fields order on the same depth (thanks @bhunjadi)
+- `getAllowedValues` now returns `null` when `allowedValues` isn't set (thanks @MohammedEssehemy)
+- Update Mocha and other dependencies
+- Readme updates (thanks @ozzywalsh)
 
 ## 1.5.3